Specific restructuring of files just to work with juceSynthFramework assumptions
The juceSynthFramework project wouldn't compile in Xcode on macOS without moving some files around, so this is a special fork just for that purpose, and I'll update the README accordingly. It involved less changes overall than finding every unfortunate hard-wired file and directory path in that JUCE project, most of which don't match what is described in its README. Possibly the Maximilian library's structure changed after those comments were written. Anyway, in order not to break other projects, I made copies of the not-found files vs. moving them outright. Maybe not a Best Practice, but this is hopefully just a short-term hack until the JUCE project is updated further to be completely up-to-date with its tools and dependencies.